third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/corner-shape-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property corner-shape value 'round'
  assert_true: corner-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner-shape value 'angle'
  assert_true: corner-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner-shape value 'round angle'
  assert_true: corner-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner-shape value 'round angle round'
  assert_true: corner-shape doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner-shape value 'round angle round angle'
  assert_true: corner-shape doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
